+++ .eslintrc.js
@@ -1,3 +1,4 @@
+
 module.exports = {
     extends: ['airbnb', 'prettier'],
     plugins: ['prettier'],
