+++ apps/innovation-ideas-react/.gitignore
