+++ apps/innovation-ideas-react/proxy/routes/index.js
@@ -1,0 +1,7 @@
+const { Router } = require('express')
+
+const router = Router()
+
+router.use('/', (req, res) => res.json({ hello: 'Hola mundo' }))
+
+module.exports = router
