+++ apps/innovation-ideas-react/public/favicon.ico
