+++ apps/innovation-ideas-react/public/manifest.json
