+++ apps/innovation-ideas-react/src/index.css
