+++ lerna.json
@@ -1,7 +1,6 @@
 {
     "lerna": "2.11.0",
     "packages": [
-        "services/*",
         "packages/*",
         "apps/*"
     ],
