+++ package.json
@@ -18,29 +18,26 @@
     },
     "homepage": "https://github.com/foxdieCR/innovation-ideas#readme",
     "dependencies": {
-        "lerna": "^2.11.0"
+        "lerna": "^3.10.7"
     },
     "devDependencies": {
+        "babel-eslint": "9.0.0",
         "cz-conventional-changelog-emoji": "^0.1.0",
-        "eslint": "^4.19.1",
-        "eslint-config-airbnb-base": "^12.1.0",
-        "eslint-config-prettier": "^2.9.0",
-        "eslint-plugin-import": "^2.11.0",
-        "eslint-plugin-prettier": "^2.6.0",
-        "prettier": "^1.12.1",
-        "babel-eslint": "^8.2.6",
-        "eslint-config-airbnb": "^17.0.0",
-        "eslint-plugin-jsx-a11y": "^6.1.1",
-        "eslint-plugin-react": "^7.10.0",
-        "lerna": "^3.0.0-rc.0"
+        "eslint": "5.6.0",
+        "eslint-config-airbnb": "^17.1.0",
+        "eslint-config-prettier": "^4.0.0",
+        "eslint-plugin-import": "^2.16.0",
+        "eslint-plugin-jsx-a11y": "^6.2.0",
+        "eslint-plugin-prettier": "^3.0.1",
+        "eslint-plugin-react": "^7.12.4",
+        "prettier": "^1.16.3"
     },
     "config": {
         "commitizen": {
-        "path": "./node_modules/cz-conventional-changelog-emoji"
+            "path": "./node_modules/cz-conventional-changelog-emoji"
         }
     },
     "workspaces": [
-        "services/*",
         "packages/*",
         "apps/*"
     ]
