--- services/api/README.md
@@ -1,4 +1,0 @@
-#Innovation-ideas API
-Api
-
-## TODO: fix README
