--- services/api/server.js
@@ -1,3 +1,0 @@
-const api = require('./server/config/app')
-
-api.start()
