--- services/api/server/config/app.js
@@ -1,5 +1,0 @@
-const express = require('./express')
-
-module.exports.start = function start() {
-  express.init()
-}
