--- services/api/server/routes/auth.route.js
@@ -1,9 +1,0 @@
-const express = require('express')
-
-const authController = require('../controllers/Auth_Controller')
-
-const router = express.Router()
-
-router.route('/').get(authController.test)
-
-module.exports = router
