--- services/api/server/routes/index.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const authRoutes = require('./auth.route')
-
-const router = express.Router()
-
-router.use('/auth', authRoutes)
-
-module.exports = router
